Criando Seeder User

+++ database/migrations/2022_03_20_191507_create_companies_table.php
@@ -18,7 +18,7 @@ class CreateCompaniesTable extends Migration
 
             $table->string('name');
             $table->string('email');
-            $table->integer('cnpj')->unique();
+            $table->bigInteger('cnpj')->unique();
             $table->tinyInteger('status')->default(1);
             $table->timestamps();
         });
